requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==19.3.0
Automat==0.8.0
blinker==1.4
certifi==2019.11.28
chardet==3.0.4
charset-normalizer==3.0.1
Click==7.0
cloud-init==22.3.4
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
datasets==2.10.0
dbus-python==1.2.16
dill==0.3.6
distro==1.4.0
distro-info===0.23ubuntu1
ec2-hibinit-agent==1.0.0
entrypoints==0.3
filelock==3.9.0
frozenlist==1.3.3
fsspec==2023.1.0
hibagent==1.0.1
httplib2==0.14.0
huggingface-hub==0.12.1
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
incremental==16.10.1
Jinja2==2.10.1
joblib==1.2.0
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
MarkupSafe==1.1.0
more-itertools==4.2.0
multidict==6.0.4
multiprocess==0.70.14
netifaces==0.10.4
np==1.0.2
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.1.0
openai==0.26.5
packaging==23.0
pandas==1.5.3
pexpect==4.6.0
pyarrow==11.0.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.0+ubuntu0.20.4.8
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
pytz==2022.7.1
PyYAML==5.3.1
regex==2022.10.31
requests==2.22.0
requests-unixsocket==0.2.0
responses==0.18.0
scikit-learn==1.2.1
scipy==1.10.1
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
sklearn==0.0.post1
sos==4.4
ssh-import-id==5.10
systemd-python==234
threadpoolctl==3.1.0
tokenizers==0.13.2
torch==1.13.1
tqdm==4.64.1
transformers==4.26.1
Twisted==18.9.0
typing-extensions==4.5.0
ubuntu-advantage-tools==27.11.3
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.14
wadllib==1.3.3
xxhash==3.2.0
yarl==1.8.2
zipp==1.0.0
zope.interface==4.7.1
